+++ R/checkValue.R
@@ -5,7 +5,7 @@
 #' Function to look for occurrences of a specific value in a \code{GADSdat}.
 #'
 #' The function checks occurrences of a specific value in a set of variables (default: all variables) in the \code{GADSdat} and outputs a vector
-#' containing the count of occurrences for all variables in which the value occurs.
+#' containing the count of occurrences for all variables in which the value occurs. It explicitly supports checking for \code{NA}.
 #'
 #'@param GADSdat \code{GADSdat} object imported via \code{eatGADS}.
 #'@param value Single string indicating how missing labels are commonly named in the value labels.
@@ -36,7 +36,8 @@ checkValue.GADSdat <- function(GADSdat, value, vars = namesGADS(GADSdat)) {
   names(name_indicator) <- vars
 
   for(nam in vars) {
-    name_indicator[nam] <- length(which(GADSdat[["dat"]][, nam] == value))
+    #name_indicator[nam] <- length(which(GADSdat[["dat"]][, nam] == value))
+    name_indicator[nam] <- sum(match(GADSdat[["dat"]][, nam], value), na.rm = TRUE)
   }
 
   name_indicator_out <- name_indicator[name_indicator > 0]